chore: remove old seek module in favor of new
