+++ nomt/src/lib.rs
@@ -38,8 +38,6 @@ mod rw_pass_cell;
 mod seek;
 mod store;
 
-mod new_seek;
-
 mod io;
 
 const MAX_COMMIT_CONCURRENCY: usize = 64;
